msm: audio: qdsp6v2: Fix voice RTAC
There was data type mismatch which caused the corruption during
memory copy. It leads to the voice RTAC failure.

CRs-Fixed: 391392

Conflicts:

	arch/arm/mach-msm/qdsp6v2/rtac.c

(cherry picked from commit 94c498ec3380730a511a888104eb3ead12161967)

+++ arch/arm/mach-msm/qdsp6v2/rtac.c
@@ -752,12 +752,12 @@ done:
 
 u32 send_voice_apr(u32 mode, void *buf, u32 opcode)
 {
-	s32				result;
-	u32				count = 0;
-	u32				bytes_returned = 0;
-	u32				payload_size;
-	u16				dest_port;
-	struct apr_hdr			voice_params;
+	s32	result;
+	u32	count = 0;
+	u32	bytes_returned = 0;
+	u32	payload_size;
+	u32	dest_port;
+	struct	apr_hdr	voice_params;
 	pr_debug("%s\n", __func__);
 
 	if (copy_from_user(&count, (void *)buf, sizeof(count))) {
